requirements.txt
flask==2.0.1
flask_socketio==5.1.0
flask_login==0.5.0
flask_pymongo==2.3.0
python_dotenv==0.18.0
pymongo[srv]==3.11.4
pytest==6.2.4
pytest-mongodb==2.2.0
pytest-dotenv==0.5.2
pytest-cov==2.12.1
bcrypt==3.2.0